Notifications ajouter / supprimer fonctionnel

+++ PROJETJSAvancé/ProjetDepart/js/testLocalStorage.js
@@ -1,39 +1,39 @@
 
 let recupStringPanierStorage = localStorage.getItem('panierStorage');
-let recupPanierStorage = JSON.parse(recupStringPanierStorage)   
-console.log(recupPanierStorage)
+// console.log(recupStringPanierStorage);
+let recupPanierStorage = JSON.parse(recupStringPanierStorage)
+// console.log("local storage = ", recupPanierStorage);
 
-for (let i = 0; i < recupPanierStorage.length; i++) {
+if (recupPanierStorage != null) {
 
-        const carte = document.createElement('tr');
-        let carteStorage = recupPanierStorage[i];
-        console.log(carteStorage)
+        for (let i = 0; i < recupPanierStorage.length; i++) {
 
+                const carte = document.createElement('tr');
+                let carteStorage = recupPanierStorage[i];
+                // console.log(carteStorage)
 
-        const img = document.createElement('td');
-        const test = carteStorage.img;
-        img.innerHTML += `<img src=${test}>`;
-        carte.appendChild(img);
+                const img = document.createElement('td');
+                const test = carteStorage.img;
+                img.innerHTML += `<img src=${test}>`;
+                carte.appendChild(img);
 
-        const nom = document.createElement('td');
-        nom.innerHTML = carteStorage.nom;
-        carte.appendChild(nom);
+                const nom = document.createElement('td');
+                nom.innerHTML = carteStorage.nom;
+                carte.appendChild(nom);
 
-        const prix = document.createElement('td');
-        prix.innerHTML = carteStorage.prix;
-        carte.appendChild(prix);
+                const prix = document.createElement('td');
+                prix.innerHTML = carteStorage.prix;
+                carte.appendChild(prix);
 
-        const quantite = document.createElement('td');
-        quantite.innerHTML = carteStorage.quantite;
-        carte.appendChild(quantite);
+                const quantite = document.createElement('td');
+                quantite.innerHTML = carteStorage.quantite;
+                carte.appendChild(quantite);
 
-        const supprimer = document.createElement('td');
-        supprimer.innerHTML = carteStorage.supprimer;
-        supprimer.addEventListener('click', deleteItem);
-        carte.appendChild(supprimer);
+                const supprimer = document.createElement('td');
+                supprimer.innerHTML = carteStorage.supprimer;
+                supprimer.addEventListener('click', deleteItem);
+                carte.appendChild(supprimer);
 
-        panier.appendChild(carte);
-
+                panier.appendChild(carte);
+        }
 }
-
-
